Re-commit API Token, now using RequestHeader

+++ src/main/java/Server/controller/AlbumSingerController.java
@@ -1,5 +1,6 @@
 package Server.controller;
 
+import Server.model.DAO.APIAccountDAO;
 import Server.model.DAO.AlbumSingerDAO;
 import Server.model.DB.AlbumSingerEntity;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -10,25 +11,33 @@ import org.springframework.web.bind.annotation.*;
 @RequestMapping("api/MusicSite/AlbumSinger")
 @RestController
 public class AlbumSingerController {
-        AlbumSingerDAO albumSingerDAO = new AlbumSingerDAO();
-        @RequestMapping(value = "/Post/",
-                method = RequestMethod.POST)
-        @ResponseBody
-        public ResponseEntity<?> registerActor(@RequestBody AlbumSingerEntity albumSingerEntity){
-            albumSingerDAO.save(albumSingerEntity);
-            return new ResponseEntity<>("Post completed", HttpStatus.CREATED);
+    AlbumSingerDAO albumSingerDAO = new AlbumSingerDAO();
+    APIAccountDAO apiAccountDAO = new APIAccountDAO();
+
+    @RequestMapping(value = "/Post/",
+            method = RequestMethod.POST)
+    @ResponseBody
+    public ResponseEntity<?> registerActor(@RequestHeader("apiToken") String apiToken, @RequestBody AlbumSingerEntity albumSingerEntity) {
+        if (apiToken == null || apiToken.isEmpty() || apiAccountDAO.checkToken(apiToken) == 0) {
+            return new ResponseEntity<>("Token is not valid.", HttpStatus.FORBIDDEN);
         }
-        @RequestMapping(value = "/Delete/{id}",
-                method = RequestMethod.DELETE
-        )
-        @ResponseBody
-        public ResponseEntity<?> deleteActor(@PathVariable("id") Long id){
-            if(albumSingerDAO.getByID(id)!=null){
-                albumSingerDAO.delete(id);
-                return new ResponseEntity<>("Delete Completed",HttpStatus.OK);
-            }
-            else return  new ResponseEntity<>("Delte Fail",HttpStatus.BAD_REQUEST);
+        albumSingerDAO.save(albumSingerEntity);
+        return new ResponseEntity<>("Post completed", HttpStatus.CREATED);
+    }
+
+    @RequestMapping(value = "/Delete/{id}",
+            method = RequestMethod.DELETE
+    )
+    @ResponseBody
+    public ResponseEntity<?> deleteActor(@RequestHeader("apiToken") String apiToken, @PathVariable("id") Long id) {
+        if (apiToken == null || apiToken.isEmpty() || apiAccountDAO.checkToken(apiToken) == 0) {
+            return new ResponseEntity<>("Token is not valid.", HttpStatus.FORBIDDEN);
         }
+        if (albumSingerDAO.getByID(id) != null) {
+            albumSingerDAO.delete(id);
+            return new ResponseEntity<>("Delete Completed", HttpStatus.OK);
+        } else return new ResponseEntity<>("Delete Fail", HttpStatus.BAD_REQUEST);
+    }
 
 
 }
