+++ src/main/java/Server/controller/AlbumSongController.java
@@ -1,5 +1,6 @@
 package Server.controller;
 
+import Server.model.DAO.APIAccountDAO;
 import Server.model.DAO.AlbumSingerDAO;
 import Server.model.DB.AlbumSingerEntity;
 import org.springframework.http.HttpStatus;
@@ -9,10 +10,14 @@ import org.springframework.web.bind.annotation.*;
 @RestController
 public class AlbumSongController {
     AlbumSingerDAO albumSingerDAO = new AlbumSingerDAO();
+    APIAccountDAO apiAccountDAO = new APIAccountDAO();
     @RequestMapping(value = "/Post/",
             method = RequestMethod.POST)
     @ResponseBody
-    public ResponseEntity<?> registerActor(@RequestBody AlbumSingerEntity albumSingerEntity){
+    public ResponseEntity<?> registerActor(@RequestHeader("apiToken") String apiToken, @RequestBody AlbumSingerEntity albumSingerEntity){
+        if (apiToken == null || apiToken.isEmpty() || apiAccountDAO.checkToken(apiToken) == 0) {
+            return new ResponseEntity<>("Token is not valid.", HttpStatus.FORBIDDEN);
+        }
         albumSingerDAO.save(albumSingerEntity);
         return new ResponseEntity<>("Post completed", HttpStatus.CREATED);
     }
@@ -20,7 +25,10 @@ public class AlbumSongController {
             method = RequestMethod.DELETE
     )
     @ResponseBody
-    public ResponseEntity<?> deleteActor(@PathVariable("id") Long id){
+    public ResponseEntity<?> deleteActor(@RequestHeader("apiToken") String apiToken, @PathVariable("id") Long id){
+        if (apiToken == null || apiToken.isEmpty() || apiAccountDAO.checkToken(apiToken) == 0) {
+            return new ResponseEntity<>("Token is not valid.", HttpStatus.FORBIDDEN);
+        }
         if(albumSingerDAO.getByID(id)!=null){
             albumSingerDAO.delete(id);
             return new ResponseEntity<>("Delete Completed",HttpStatus.OK);
