+++ src/main/java/Server/controller/FilmActorController.java
@@ -1,5 +1,6 @@
 package Server.controller;
 
+import Server.model.DAO.APIAccountDAO;
 import Server.model.DAO.FilmActorDAO;
 import Server.model.DB.DirectorFilmEntity;
 import Server.model.DB.FilmActorEntity;
@@ -11,22 +12,30 @@ import org.springframework.web.bind.annotation.*;
 @RestController
 public class FilmActorController {
     FilmActorDAO filmActorDAO = new FilmActorDAO();
+    APIAccountDAO apiAccountDAO = new APIAccountDAO();
+
     @RequestMapping(value = "/Post/",
             method = RequestMethod.POST)
     @ResponseBody
-    public ResponseEntity<?> registerActor(@RequestBody FilmActorEntity directorFilmEntity){
+    public ResponseEntity<?> registerActor(@RequestHeader("apiToken") String apiToken, @RequestBody FilmActorEntity directorFilmEntity) {
+        if (apiToken == null || apiToken.isEmpty() || apiAccountDAO.checkToken(apiToken) == 0) {
+            return new ResponseEntity<>("Token is not valid.", HttpStatus.FORBIDDEN);
+        }
         filmActorDAO.save(directorFilmEntity);
         return new ResponseEntity<>("Post completed", HttpStatus.CREATED);
     }
+
     @RequestMapping(value = "/Delete/{id}",
             method = RequestMethod.DELETE
     )
     @ResponseBody
-    public ResponseEntity<?> deleteActor(@PathVariable("id") Long id){
-        if(filmActorDAO.getByID(id)!=null){
-            filmActorDAO.delete(id);
-            return new ResponseEntity<>("Delete Completed",HttpStatus.OK);
+    public ResponseEntity<?> deleteActor(@RequestHeader("apiToken") String apiToken, @PathVariable("id") Long id) {
+        if (apiToken == null || apiToken.isEmpty() || apiAccountDAO.checkToken(apiToken) == 0) {
+            return new ResponseEntity<>("Token is not valid.", HttpStatus.FORBIDDEN);
         }
-        else return  new ResponseEntity<>("Delte Fail",HttpStatus.BAD_REQUEST);
+        if (filmActorDAO.getByID(id) != null) {
+            filmActorDAO.delete(id);
+            return new ResponseEntity<>("Delete Completed", HttpStatus.OK);
+        } else return new ResponseEntity<>("Delte Fail", HttpStatus.BAD_REQUEST);
     }
 }
