+++ src/main/java/Server/controller/SerifilmFilmController.java
@@ -1,5 +1,6 @@
 package Server.controller;
 
+import Server.model.DAO.APIAccountDAO;
 import Server.model.DAO.SerifilmFilmDAO;
 import Server.model.DB.SerifilmFilmEntity;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -12,22 +13,30 @@ import org.springframework.web.bind.annotation.*;
 public class SerifilmFilmController {
     @Autowired
     SerifilmFilmDAO serifilmFilmDAO;
+    APIAccountDAO apiAccountDAO = new APIAccountDAO();
+
     @RequestMapping(value = "/Post/",
             method = RequestMethod.POST)
     @ResponseBody
-    public ResponseEntity<?> registerActor(@RequestBody SerifilmFilmEntity serifilmFilmEntity){
+    public ResponseEntity<?> registerActor(@RequestHeader("apiToken") String apiToken, @RequestBody SerifilmFilmEntity serifilmFilmEntity) {
+        if (apiToken == null || apiToken.isEmpty() || apiAccountDAO.checkToken(apiToken) == 0) {
+            return new ResponseEntity<>("Token is not valid.", HttpStatus.FORBIDDEN);
+        }
         serifilmFilmDAO.save(serifilmFilmEntity);
         return new ResponseEntity<>("Post completed", HttpStatus.CREATED);
     }
+
     @RequestMapping(value = "/Delete/{id}",
             method = RequestMethod.DELETE
     )
     @ResponseBody
-    public ResponseEntity<?> deleteActor(@PathVariable("id") Long id){
-        if(serifilmFilmDAO.getByID(id)!=null){
-            serifilmFilmDAO.delete(id);
-            return new ResponseEntity<>("Delete Completed",HttpStatus.OK);
+    public ResponseEntity<?> deleteActor(@RequestHeader("apiToken") String apiToken, @PathVariable("id") Long id) {
+        if (apiToken == null || apiToken.isEmpty() || apiAccountDAO.checkToken(apiToken) == 0) {
+            return new ResponseEntity<>("Token is not valid.", HttpStatus.FORBIDDEN);
         }
-        else return  new ResponseEntity<>("Delte Fail",HttpStatus.BAD_REQUEST);
+        if (serifilmFilmDAO.getByID(id) != null) {
+            serifilmFilmDAO.delete(id);
+            return new ResponseEntity<>("Delete Completed", HttpStatus.OK);
+        } else return new ResponseEntity<>("Delte Fail", HttpStatus.BAD_REQUEST);
     }
 }
