+++ src/main/java/Server/controller/SongCategorySongController.java
@@ -1,5 +1,6 @@
 package Server.controller;
 
+import Server.model.DAO.APIAccountDAO;
 import Server.model.DAO.SongCategorySongDAO;
 import Server.model.DB.SongCategorysongEntity;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -12,10 +13,14 @@ import org.springframework.web.bind.annotation.*;
 public class SongCategorySongController {
     @Autowired
     SongCategorySongDAO songCategorySongDAO;
+    APIAccountDAO apiAccountDAO = new APIAccountDAO();
     @RequestMapping(value = "/Post/",
             method = RequestMethod.POST)
     @ResponseBody
-    public ResponseEntity<?> registerActor(@RequestBody SongCategorysongEntity serifilmFilmEntity){
+    public ResponseEntity<?> registerActor(@RequestHeader("apiToken") String apiToken, @RequestBody SongCategorysongEntity serifilmFilmEntity){
+        if (apiToken == null || apiToken.isEmpty() || apiAccountDAO.checkToken(apiToken) == 0) {
+            return new ResponseEntity<>("Token is not valid.", HttpStatus.FORBIDDEN);
+        }
         songCategorySongDAO.save(serifilmFilmEntity);
         return new ResponseEntity<>("Post completed", HttpStatus.CREATED);
     }
@@ -23,7 +28,10 @@ public class SongCategorySongController {
             method = RequestMethod.DELETE
     )
     @ResponseBody
-    public ResponseEntity<?> deleteActor(@PathVariable("id") Long id){
+    public ResponseEntity<?> deleteActor(@RequestHeader("apiToken") String apiToken, @PathVariable("id") Long id){
+        if (apiToken == null || apiToken.isEmpty() || apiAccountDAO.checkToken(apiToken) == 0) {
+            return new ResponseEntity<>("Token is not valid.", HttpStatus.FORBIDDEN);
+        }
         if(songCategorySongDAO.getByID(id)!=null){
             songCategorySongDAO.delete(id);
             return new ResponseEntity<>("Delete Completed",HttpStatus.OK);
