+++ src/main/java/Server/controller/SongSingerController.java
@@ -1,6 +1,7 @@
 package Server.controller;
 
 
+import Server.model.DAO.APIAccountDAO;
 import Server.model.DAO.SongSingerDAO;
 import Server.model.DB.SongSingerEntity;
 import org.springframework.http.HttpStatus;
@@ -11,22 +12,30 @@ import org.springframework.web.bind.annotation.*;
 @RestController
 public class SongSingerController {
     SongSingerDAO songSingerDAO = new SongSingerDAO();
+    APIAccountDAO apiAccountDAO = new APIAccountDAO();
+
     @RequestMapping(value = "/Post/",
             method = RequestMethod.POST)
     @ResponseBody
-    public ResponseEntity<?> registerActor(@RequestBody SongSingerEntity albumSingerEntity){
+    public ResponseEntity<?> registerActor(@RequestHeader("apiToken") String apiToken, @RequestBody SongSingerEntity albumSingerEntity) {
+        if (apiToken == null || apiToken.isEmpty() || apiAccountDAO.checkToken(apiToken) == 0) {
+            return new ResponseEntity<>("Token is not valid.", HttpStatus.FORBIDDEN);
+        }
         songSingerDAO.save(albumSingerEntity);
         return new ResponseEntity<>("Post completed", HttpStatus.CREATED);
     }
+
     @RequestMapping(value = "/Delete/{id}",
             method = RequestMethod.DELETE
     )
     @ResponseBody
-    public ResponseEntity<?> deleteActor(@PathVariable("id") Long id){
-        if(songSingerDAO.getByID(id)!=null){
-            songSingerDAO.delete(id);
-            return new ResponseEntity<>("Delete Completed",HttpStatus.OK);
+    public ResponseEntity<?> deleteActor(@RequestHeader("apiToken") String apiToken, @PathVariable("id") Long id) {
+        if (apiToken == null || apiToken.isEmpty() || apiAccountDAO.checkToken(apiToken) == 0) {
+            return new ResponseEntity<>("Token is not valid.", HttpStatus.FORBIDDEN);
         }
-        else return  new ResponseEntity<>("Delte Fail",HttpStatus.BAD_REQUEST);
+        if (songSingerDAO.getByID(id) != null) {
+            songSingerDAO.delete(id);
+            return new ResponseEntity<>("Delete Completed", HttpStatus.OK);
+        } else return new ResponseEntity<>("Delte Fail", HttpStatus.BAD_REQUEST);
     }
 }
